requirements.txt
Flask==1.1.1
requests
SQLAlchemy==1.3.12
Flask-SQLAlchemy
pg8000
psycopg2-binary==2.8.4
gunicorn
flask_migrate
google-cloud-datastore
simplejson
flask-marshmallow
flask-jwt-extended
